Add player field view

+++ src/Helpers/Size.java
@@ -2,27 +2,27 @@ package Helpers;
 
 public class Size {
 
-	private float width;
-	private float height;
+	private int width;
+	private int height;
 	
-	public Size(float width, float height) {
+	public Size(int width, int height) {
 		setWidth(width);
 		setHeight(height);
 	}
 	
-	public float getWidth() {
+	public int getWidth() {
 		return width;
 	}
 	
-	public void setWidth(float width) {
+	public void setWidth(int width) {
 		this.width = width;
 	}
 	
-	public float getHeight() {
+	public int getHeight() {
 		return height;
 	}
 	
-	public void setHeight(float height) {
+	public void setHeight(int height) {
 		this.height = height;
 	}
